requirements.txt
pandas
numpy
sqlalchemy
psycopg2-binary
python-dotenv
requests
statsmodels
mlflow
scikit-learn
pmdarima